feat: extract dashboard page to separate component
